pyproject.toml
[project]
name = "e"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.4",
    "python-dotenv>=1.0.1",
    "questionary>=2.1.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
